// frontend/src/types/pointSettings.ts
export type PointSettings = {
  FIRST_LOGIN_BONUS: number;
  BLOG_REWARD_POINTS: number;
  DAILY_BLOG_REWARD_LIMIT: number;
  PREDICTION_WINNER_ENTRY_COST: number;
  PREDICTION_TOP3_ENTRY_COST: number;
  PREDICTION_WINNER_REWARD: number;
  PREDICTION_TOP3_EXACT_REWARD: number;
  PREDICTION_TOP3_ANY_ORDER_REWARD: number;
};